Added a unit-test which ensures that the --noout parameter can actually write to its output (/dev/null).
This adds a new basic test that confirms that the --noout parameter is
actually writing to "/dev/null". Since we can't split stdout and stderr
with bats due to the way it's been configured, we're explicitly checking
that running podman with no parameters emits just the missing command
error rather than both golang's error message of being unable to write to
"/dev/null" and the missing command error.

Signed-off-by: Ali Rizvi-Santiago <arizvisa@gmail.com>