Added some more unit-tests for the --out parameter which verify that pause, stop, pod, and network will echo their exact parameter.
This checks the output of podman-stop(1), podman-pause(1), podman-pod-create(1),
and podman-network-create(1) using the new --out parameter. Each of these
will actually just echo whatever was given on the command-line if the call
succeeds. So to actually test that the output is correct, we capture the output
that is echo'd, convert it to an id, and then confirm that the id exists

Signed-off-by: Ali Rizvi-Santiago <arizvisa@gmail.com>